upd-enerhy

+++ src/structs/vec.tpp
@@ -34,7 +34,7 @@ T Vec<T, N, R>::get_coord(size_t index) {
 template VECTORTEMPLATE
 Vec<T, N, R> Vec<T, N, R>::filled(T val) { 
     Vec<T, N, R> v;
-    for (int i =0; i<N; i++) v.set_coord(i, val);
+    for (size_t i =0; i<N; i++) v.set_coord(i, val);
     return v;  
 }
 
@@ -43,7 +43,7 @@ template VECTORTEMPLATE
 T Vec<T, N, R>::length(){
     T r = 0;
 
-    for(int i=0; i<N; i++){
+    for(size_t i=0; i<N; i++){
         r += this->dat[i]*this->dat[i];
     }
 
@@ -53,7 +53,7 @@ T Vec<T, N, R>::length(){
 template VECTORTEMPLATE
 T Vec<T, N, R>::sqlength(){
     T r = 0;
-    for(int i=0; i<N; i++){
+    for(size_t i=0; i<N; i++){
         r += this->dat[i]*this->dat[i];
     }
     return r;
